+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `compas_occ.brep.Brep.from_plane`.
+* Added `compas_occ.brep.Brep.simplify`.
+
 ### Changed
 
 ### Removed